Added serialised fields to hold weapon atoms for selection/generation
fixed small issue with sword_hilt_0 having an additional weapon component. This was causing issues with the item info window rendering the wrong (null) data.

Small changes to the GrammarEngine which prevent hereditary scaling problems when making an object a child. Also added methods for invoking grammar rewriting on a specific object.